requirements.txt
-e git://github.com/django/django.git@8568e7cfa4fb87cd0a9ead1b3fbb7a39d19e98b9#egg=django
Markdown==2.4.1
boto==2.32.1
dj-database-url==0.3.0
dj-static==0.0.6
django-endless-pagination==2.0
django-markdown==0.6.1
django-storages==1.1.8
django-taggit==0.12
django-toolbelt==0.0.1
docutils==0.12
gunicorn==19.0.0
psycopg2==2.5.3
static3==0.5.1
waitress==0.8.9
